+++ lib/context/context_app.dart
@@ -3,7 +3,7 @@ library context_app;
 
 import 'package:package_info/package_info.dart' show PackageInfo;
 
-import '../util/util.dart' show debugError;
+import '../debug/debug.dart' show Debug;
 
 /// @nodoc
 Future<Map<String, dynamic>> contextApp() async {
@@ -17,7 +17,7 @@ Future<Map<String, dynamic>> contextApp() async {
       'version': packageInfo.version
     };
   } catch (e, s) {
-    debugError(e, s);
+    Debug().error(e, s);
 
     return <String, dynamic>{};
   }