+++ lib/util/util.dart
@@ -1,22 +1,11 @@
 /// @nodoc
 library util;
 
-import 'package:flutter/foundation.dart' show debugPrint;
 import 'package:recase/recase.dart' show ReCase;
 
 /// @nodoc
 String dartEnv() =>
     bool.fromEnvironment('dart.vm.product') ? 'PRODUCTION' : 'DEVELOPMENT';
-
-/// @nodoc
-void debugError(dynamic e, [dynamic s]) => _debug(() {
-      debugPrint('flutter_analytics ➲ ERROR:\n$e${s == null ? '' : '\n$s'}\n');
-    });
-
-/// @nodoc
-void debugLog(dynamic msg) => _debug(() {
-      debugPrint('flutter_analytics ➲ ${msg.toString()} @ ${_isoNow()}\n');
-    });
 
 /// @nodoc
 String camelCase(String s) => _toCase(s, (s) => ReCase(s).camelCase);
@@ -35,16 +24,3 @@ String _toCase(String string, String Function(String) recaseFunction) {
     return null;
   }
 }
-
-void _debug(void Function() exec) {
-  if (!bool.fromEnvironment('dart.vm.product')) {
-    exec();
-  }
-}
-
-String _isoNow() {
-  return DateTime.now().toUtc().toIso8601String();
-}
-
-/// @nodoc
-typedef OnEvent<T> = Future<void> Function(T);